Bug #1913083 soc: sprd: Coverity fix
[root cause  ] Coverity fix
[changes     ] Coverity fix
[side effects] NA
[self test   ]:Yes
[download normally]:Yes
[power on/off normally]:Yes
[do common repository/branch inspection]:Yes
[is there dependence]:No
[confirm dependent commit]:Yes
[board]:a12
[test case]:boot
[reviewers   ] ling_ling.xu
[change_type ] feature
[tag_product ] common

+++ drivers/soc/sprd/power/power_debug/sprd_pdbg.c
@@ -553,7 +553,7 @@ static int sprd_pdbg_thread(void *data)
 		SPRD_PDBG_INFO("#---------PDBG LIGHT SLEEP END-----------#\n");
 
 		set_current_state(TASK_INTERRUPTIBLE);
-		schedule_timeout(pdbg->scan_interval * HZ);
+		schedule_timeout(pdbg->scan_interval * (long)HZ);
 	}
 
 	return 0;
